Add support for --force and --non-interactive flag
This commit adds support for -f or --force to indicate that the script has permissions to `rm -rf` the provided $APP directory. If the flag is not provided, the user is prompted whether or or not they would like to overwrite the existing directory.

It also adds support for the -n or --non-interactive flag. This indicates whether or not user interaction should be required for the script; if this is set, then the default options for the script will be executed.